refactor(src): model aufteilung vorgenommen
