+++ src/game/StateTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import questions.Category;
-import questions.Question;
+import model.Category;
+import model.Question;
 
 /**
  * Klasse zum testen des Spiels
