+++ src/questions/QuestionParser.java
@@ -1,20 +1,17 @@
 package questions;
 
 import java.io.BufferedReader;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.InputStream;
+
 import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
-import java.io.PrintWriter;
-import java.io.StringWriter;
-import java.nio.charset.Charset;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.stream.Collectors;
 
 import com.google.gson.*;
+
+import model.Question;
+import model.Category;
+
 
 public class QuestionParser implements QuestionStore {
